+++ src/data/store/arena/string/impls/mod.rs
@@ -1,0 +1,4 @@
+// devela/src/data/store/arena/string/impls/mod.rs
+
+mod array; // __arena_string_impl_array!
+mod vec; // __arena_string_impl_vec!